// src/generated/graphql.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
export type RequireFields<T, K extends keyof T> = { [X in Exclude<keyof T, K>]?: T[X] } & { [P in K]-?: NonNullable<T[P]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
};

export type Mutation = {
  __typename?: 'Mutation';
  createToDo?: Maybe<ToDo>;
  updateToDo?: Maybe<ToDo>;
  deleteToDo?: Maybe<ToDo>;
  deleteToDos: Array<ToDo>;
};


export type MutationCreateToDoArgs = {
  toDoInput?: Maybe<ToDoInput>;
};


export type MutationUpdateToDoArgs = {
  toDoId: Scalars['ID'];
  toDoInput?: Maybe<ToDoInput>;
};


export type MutationDeleteToDoArgs = {
  toDoId: Scalars['ID'];
};

export type Query = {
  __typename?: 'Query';
  getToDo: ToDo;
  getToDos: Array<ToDo>;
};


export type QueryGetToDoArgs = {
  toDoId: Scalars['ID'];
};

export type ToDo = {
  __typename?: 'ToDo';
  id: Scalars['ID'];
  title: Scalars['String'];
  description: Scalars['String'];
};

export type ToDoInput = {
  title: Scalars['String'];
  description: Scalars['String'];
};



export type ResolverTypeWrapper<T> = Promise<T> | T;


export type ResolverWithResolve<TResult, TParent, TContext, TArgs> = {
  resolve: ResolverFn<TResult, TParent, TContext, TArgs>;
};
export type Resolver<TResult, TParent = {}, TContext = {}, TArgs = {}> = ResolverFn<TResult, TParent, TContext, TArgs> | ResolverWithResolve<TResult, TParent, TContext, TArgs>;
// ...
export type IsTypeOfResolverFn<T = {}, TContext = {}> = (obj: T, context: TContext, info: GraphQLResolveInfo) => boolean | Promise<boolean>;

export type NextResolverFn<T> = () => Promise<T>;

export type DirectiveResolverFn<TResult = {}, TParent = {}, TContext = {}, TArgs = {}> = (
  next: NextResolverFn<TResult>,
  parent: TParent,
  args: TArgs,
  context: TContext,
  info: GraphQLResolveInfo
) => TResult | Promise<TResult>;

/** Mapping between all available schema types and the resolvers types */
export type ResolversTypes = {
  Mutation: ResolverTypeWrapper<{}>;
  ID: ResolverTypeWrapper<Scalars['ID']>;
  Query: ResolverTypeWrapper<{}>;
  ToDo: ResolverTypeWrapper<ToDo>;
  String: ResolverTypeWrapper<Scalars['String']>;
  ToDoInput: ToDoInput;
  Boolean: ResolverTypeWrapper<Scalars['Boolean']>;
};

/** Mapping between all available schema types and the resolvers parents */
export type ResolversParentTypes = {
  Mutation: {};
  ID: Scalars['ID'];
  Query: {};
  ToDo: ToDo;
  String: Scalars['String'];
  ToDoInput: ToDoInput;
  Boolean: Scalars['Boolean'];
};

export type MutationResolvers<ContextType = any, ParentType extends ResolversParentTypes['Mutation'] = ResolversParentTypes['Mutation']> = {
  createToDo?: Resolver<Maybe<ResolversTypes['ToDo']>, ParentType, ContextType, RequireFields<MutationCreateToDoArgs, never>>;
  updateToDo?: Resolver<Maybe<ResolversTypes['ToDo']>, ParentType, ContextType, RequireFields<MutationUpdateToDoArgs, 'toDoId'>>;
  deleteToDo?: Resolver<Maybe<ResolversTypes['ToDo']>, ParentType, ContextType, RequireFields<MutationDeleteToDoArgs, 'toDoId'>>;
  deleteToDos?: Resolver<Array<ResolversTypes['ToDo']>, ParentType, ContextType>;
};

export type QueryResolvers<ContextType = any, ParentType extends ResolversParentTypes['Query'] = ResolversParentTypes['Query']> = {
  getToDo?: Resolver<ResolversTypes['ToDo'], ParentType, ContextType, RequireFields<QueryGetToDoArgs, 'toDoId'>>;
  getToDos?: Resolver<Array<ResolversTypes['ToDo']>, ParentType, ContextType>;
};

export type ToDoResolvers<ContextType = any, ParentType extends ResolversParentTypes['ToDo'] = ResolversParentTypes['ToDo']> = {
  id?: Resolver<ResolversTypes['ID'], ParentType, ContextType>;
  title?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
  description?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};

export type Resolvers<ContextType = any> = {
  Mutation?: MutationResolvers<ContextType>;
  Query?: QueryResolvers<ContextType>;
  ToDo?: ToDoResolvers<ContextType>;
};